Added Carter - Error

+++ Jarvis.py
@@ -9,6 +9,7 @@ from nextcord.ext import commands
 from Utils.secret import prefix, token, carterAPI
 from Utils.Carter import CarterSend
 
+
 intents = nextcord.Intents.default()
 
 intents.message_content = True
@@ -35,26 +36,29 @@ async def on_ready():
 
 
 @bot.event
-async def on_message(message): 
-    if message.author == bot.user:
-        return
-    
-    user = message.author
-    sentence = message.content
-    sentence = sentence.lower()
-    
-    if prefix in sentence:
-        CarterSend(sentence, user.id)
-        with open('CarterResponse.txt') as f:
-            ResponseOutput = f.read()
+async def on_command_error(message, error):
+    if isinstance(error, commands.CommandNotFound):
+        if message.author == bot.user:
+            return
 
-        print(message.content)
-        await message.channel.send(f"{ResponseOutput}")
-        print(ResponseOutput)
-        os.remove("CarterResponse.txt")
-    
-    else: 
-        pass
+        user = message.author
+        sentence = str(message.content)
+        sentence = sentence.lower()
+
+        if prefix in sentence:
+            CarterSend(sentence, user.id)
+            with open('CarterResponse.txt') as f:
+                ResponseOutput = f.read()
+
+            print(message.content)
+            await message.channel.send(f"{ResponseOutput}")
+            print(ResponseOutput)
+            os.remove("CarterResponse.txt")
+
+        else: 
+            pass
+    else:
+        raise error
 
 # Run Discord bot
 bot.run(token)
